requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator